SecondCommit
Includes a socket that charges up filaman, a second level featuring this
socket, a fleshed out tutorial/first level. Invisible red ring for
platforms that can appear. Filaman turns muscly when purple, can double
jump when red. Currently controlled with 1, 2 and 3. Can press E in
certain scenarios to interact with the dialogue wall. Features scene
changes that cause the game to fade to white and load new levels when
filaman has reached the end. Has a fade through black sequence that
plays when filaman falls off the level, he is then transported to a
convenient platform.